Updated router for onload function support

+++ hosting/src/controllers/manager/menu.ts
@@ -71,7 +71,7 @@ export async function addDish(data: {
   const dish: DishDocument = await DishCollection.create(
     data.name,
     data.description,
-    parseInt(data.price)
+    parseFloat(data.price)
   );
   menu.addDish(dish);
 }
